pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.71.0",
    "cryptography>=46.0.3",
    "fastapi>=0.120.1",
    "groq>=0.33.0",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "openai>=2.6.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.11",
    "pycpfcnpj>=1.8",
    "pydantic>=2.12.3",
    "pypdf2>=3.0.1",
    "pypdfium2>=5.0.0",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "signxml>=4.2.0",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
    "xmltodict>=1.0.2",
    "zeep>=4.3.2",
]
